Moved exif functions to imagemeta/exif package
