+++ exif/tiffTagEnumerator_test.go
@@ -1,3 +1,3 @@
-package imagemeta
+package exif
 
 // TODO: write tests for ifdTagEnumerator
